pyproject.toml
[project]
name = "mrm-chat-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "channels>=4.3.1",
    "daphne>=4.2.1",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.10.1",
    "whitenoise>=6.11.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.1.1",
    "dj-database-url>=2.3.0",
    "openai>=1.0.0",
]
